Added Another Command That Scrapes Cities Also, 4/16/2024
Added Another Command That Scrapes Cities Also. Somebody wanted to know the best quality guildhalls for the thieves guild in a specific region, so I figured I would repurpose the dungeon scrapping feature of this to do similar but for the cities, and it worked out fairly well, just takes a bit of time to parse through all the buildings in the game essentially, but still useful utility to have, 4/16/2024.